--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -286,7 +286,7 @@ macro(add_pyston_test testname directory)
 endmacro()
 
 # tests         testname directory arguments
-add_pyston_test(defaults tests --order-by-mtime)
+add_pyston_test(defaults tests --order-by-mtime -t50)
 add_pyston_test(force_llvm tests -a=-n -a=-x -t50)
 if(${CMAKE_BUILD_TYPE} STREQUAL "Release")
   add_pyston_test(max_compilation_tier tests -a=-O -a=-x -t50)