```
           django_template3.py             2.7s (4)             2.5s (4)  -8.3%
                 pyxl_bench.py             2.2s (4)             2.2s (4)  -0.4%
     sqlalchemy_imperative2.py             2.6s (4)             2.6s (4)  -0.6%
                       geomean                 2.5s                 2.4s  -3.2%
```